Complete Day 73 challenges
- Allow the user to switch map modes, between the standard mode and
hybrid.
- Our app silently fails when errors occur during biometric
authentication, so add code to show those errors in an alert.
- Create another view model, this time for `EditView`. What you put in
the view model is down to you, but I would recommend leaving `dismiss`
and `onSave` in the view itself – the former uses the environment,
which can only be read by the view, and the latter doesn’t really add
anything when moved into the model.